# sql/parser: Change EvalContext to be passed by reference