--- a/sql/distsql/expr_test.go
+++ b/sql/distsql/expr_test.go
@@ -30,7 +30,7 @@ func (d testVarContainer) IndexedVarReturnType(idx int) parser.Datum {
 	return parser.TypeInt
 }
 
-func (d testVarContainer) IndexedVarEval(idx int, ctx parser.EvalContext) (parser.Datum, error) {
+func (d testVarContainer) IndexedVarEval(idx int, ctx *parser.EvalContext) (parser.Datum, error) {
 	return nil, nil
 }