--- a/sql/distsql/server.go
+++ b/sql/distsql/server.go
@@ -19,8 +19,6 @@ package distsql
 import (
 	"golang.org/x/net/context"
 
-	"gopkg.in/inf.v0"
-
 	"github.com/cockroachdb/cockroach/client"
 	"github.com/cockroachdb/cockroach/roachpb"
 	"github.com/cockroachdb/cockroach/sql/parser"
@@ -47,7 +45,6 @@ func NewServer(ctx ServerContext) *ServerImpl {
 		ctx: ctx,
 		evalCtx: parser.EvalContext{
 			ReCache: parser.NewRegexpCache(512),
-			TmpDec:  new(inf.Dec),
 		},
 	}
 	return ds