--- a/sql/group.go
+++ b/sql/group.go
@@ -79,7 +79,7 @@ func (p *planner) groupBy(n *parser.SelectClause, s *selectNode) (*groupNode, er
 			return nil, err
 		}
 
-		norm, err := p.parser.NormalizeExpr(p.evalCtx, typedExpr)
+		norm, err := p.parser.NormalizeExpr(&p.evalCtx, typedExpr)
 		if err != nil {
 			return nil, err
 		}
@@ -116,7 +116,7 @@ func (p *planner) groupBy(n *parser.SelectClause, s *selectNode) (*groupNode, er
 			return nil, err
 		}
 
-		typedHaving, err = p.parser.NormalizeExpr(p.evalCtx, typedHaving)
+		typedHaving, err = p.parser.NormalizeExpr(&p.evalCtx, typedHaving)
 		if err != nil {
 			return nil, err
 		}
@@ -375,7 +375,7 @@ func (n *groupNode) computeAggregates() error {
 		n.currentBucket = k
 
 		if n.having != nil {
-			res, err := n.having.Eval(n.planner.evalCtx)
+			res, err := n.having.Eval(&n.planner.evalCtx)
 			if err != nil {
 				return err
 			}
@@ -388,7 +388,7 @@ func (n *groupNode) computeAggregates() error {
 
 		row := make(parser.DTuple, 0, len(n.render))
 		for _, r := range n.render {
-			res, err := r.Eval(n.planner.evalCtx)
+			res, err := r.Eval(&n.planner.evalCtx)
 			if err != nil {
 				return err
 			}
@@ -709,7 +709,7 @@ func (a *aggregateFunc) TypeCheck(_ *parser.SemaContext, desired parser.Datum) (
 	return a, nil
 }
 
-func (a *aggregateFunc) Eval(ctx parser.EvalContext) (parser.Datum, error) {
+func (a *aggregateFunc) Eval(ctx *parser.EvalContext) (parser.Datum, error) {
 	// During init of the group buckets, grouped expressions (i.e. wrapped
 	// qvalues) are Eval()'ed to determine the bucket for a row, so pass these
 	// calls through to the underlying `arg` expr Eval until init is done.