--- a/sql/parser/indexed_vars.go
+++ b/sql/parser/indexed_vars.go
@@ -24,7 +24,7 @@ import (
 // IndexedVarContainer provides the implementation of TypeCheck, Eval, and
 // String for IndexedVars.
 type IndexedVarContainer interface {
-	IndexedVarEval(idx int, ctx EvalContext) (Datum, error)
+	IndexedVarEval(idx int, ctx *EvalContext) (Datum, error)
 	IndexedVarReturnType(idx int) Datum
 	IndexedVarString(idx int) string
 }
@@ -54,7 +54,7 @@ func (v *IndexedVar) TypeCheck(_ *SemaContext, desired Datum) (TypedExpr, error)
 }
 
 // Eval is part of the TypedExpr interface.
-func (v *IndexedVar) Eval(ctx EvalContext) (Datum, error) {
+func (v *IndexedVar) Eval(ctx *EvalContext) (Datum, error) {
 	return v.container.IndexedVarEval(v.Idx, ctx)
 }