--- a/sql/parser/parse.go
+++ b/sql/parser/parse.go
@@ -119,7 +119,7 @@ func TypeCheckAndRequire(expr Expr, ctx *SemaContext, required Datum, op string)
 
 // NormalizeExpr is wrapper around ctx.NormalizeExpr which avoids allocation of
 // a normalizeVisitor.
-func (p *Parser) NormalizeExpr(ctx EvalContext, typedExpr TypedExpr) (TypedExpr, error) {
+func (p *Parser) NormalizeExpr(ctx *EvalContext, typedExpr TypedExpr) (TypedExpr, error) {
 	p.normalizeVisitor = normalizeVisitor{ctx: ctx}
 	expr, _ := WalkExpr(&p.normalizeVisitor, typedExpr)
 	if err := p.normalizeVisitor.err; err != nil {