Fixes #6809.

This change switches all occurrences of passing `EvalContext` as a value
with passing it as a reference. This opens up more options around how we
can use the context. It was also such a big struct being passed around,
that passing it as a pointer instead gives a small perf boost.

```
name                    old time/op    new time/op    delta
Select1_Cockroach-4       56.1µs ± 2%    55.1µs ± 3%  -1.85%  (p=0.005 n=10+10)
Select2_Cockroach-4        828µs ± 2%     821µs ± 2%    ~     (p=0.165 n=10+10)
Select3_Cockroach-4       1.11ms ± 1%    1.10ms ± 3%    ~      (p=0.083 n=8+10)
Insert1_Cockroach-4        485µs ± 3%     477µs ± 3%    ~      (p=0.068 n=8+10)
Insert10_Cockroach-4       683µs ± 4%     671µs ± 2%    ~     (p=0.052 n=10+10)
Insert100_Cockroach-4     2.22ms ± 2%    2.19ms ± 3%    ~     (p=0.075 n=10+10)
Insert1000_Cockroach-4    16.5ms ± 2%    16.2ms ± 1%  -1.81%    (p=0.001 n=9+9)
Update1_Cockroach-4        751µs ± 4%     754µs ± 4%    ~      (p=0.780 n=9+10)
Update10_Cockroach-4      1.27ms ± 5%    1.28ms ± 2%    ~     (p=0.684 n=10+10)
Update100_Cockroach-4     5.49ms ± 6%    5.38ms ± 3%    ~      (p=0.182 n=9+10)
Update1000_Cockroach-4    43.3ms ± 5%    42.8ms ± 5%    ~     (p=0.315 n=10+10)
Delete1_Cockroach-4        830µs ± 5%     803µs ± 5%  -3.21%    (p=0.014 n=9+9)
Delete10_Cockroach-4      1.41ms ± 4%    1.38ms ± 3%  -2.60%  (p=0.029 n=10+10)
Delete100_Cockroach-4     6.60ms ± 5%    6.46ms ± 5%    ~     (p=0.247 n=10+10)
Delete1000_Cockroach-4    66.0ms ± 9%    66.5ms ± 5%    ~      (p=0.661 n=10+9)

name                    old alloc/op   new alloc/op   delta
Select1_Cockroach-4       2.80kB ± 0%    2.75kB ± 0%  -1.75%  (p=0.000 n=10+10)
Select2_Cockroach-4       98.6kB ± 0%    98.6kB ± 0%  -0.04%   (p=0.002 n=10+9)
Select3_Cockroach-4        141kB ± 0%     141kB ± 0%  -0.16%   (p=0.000 n=10+9)
Insert1_Cockroach-4       24.8kB ± 0%    24.7kB ± 0%  -0.20%   (p=0.000 n=9+10)
Insert10_Cockroach-4      69.3kB ± 0%    69.2kB ± 0%  -0.08%  (p=0.000 n=10+10)
Insert100_Cockroach-4      462kB ± 0%     462kB ± 0%    ~      (p=0.780 n=9+10)
Insert1000_Cockroach-4    4.05MB ± 0%    4.05MB ± 0%    ~      (p=0.661 n=9+10)
Update1_Cockroach-4       50.6kB ± 0%    50.5kB ± 0%  -0.10%  (p=0.000 n=10+10)
Update10_Cockroach-4       124kB ± 0%     124kB ± 0%  -0.04%    (p=0.000 n=8+9)
Update100_Cockroach-4      817kB ± 0%     817kB ± 0%  -0.01%    (p=0.000 n=9+9)
Update1000_Cockroach-4    6.88MB ± 0%    6.90MB ± 0%    ~      (p=0.274 n=8+10)
Delete1_Cockroach-4       44.0kB ± 0%    44.0kB ± 0%  -0.14%    (p=0.000 n=9+9)
Delete10_Cockroach-4      71.5kB ± 0%    71.5kB ± 0%  -0.06%   (p=0.017 n=9+10)
Delete100_Cockroach-4      336kB ± 0%     336kB ± 0%  -0.07%   (p=0.005 n=10+8)
Delete1000_Cockroach-4    3.55MB ± 2%    3.56MB ± 1%    ~     (p=0.971 n=10+10)

name                    old allocs/op  new allocs/op  delta
Select1_Cockroach-4         52.0 ± 0%      51.0 ± 0%  -1.92%  (p=0.000 n=10+10)
Select2_Cockroach-4        1.42k ± 0%     1.42k ± 0%  -0.07%   (p=0.000 n=10+8)
Select3_Cockroach-4        2.27k ± 0%     2.27k ± 0%  -0.09%   (p=0.000 n=10+9)
Insert1_Cockroach-4          269 ± 0%       267 ± 0%  -0.48%  (p=0.001 n=10+10)
Insert10_Cockroach-4         486 ± 0%       485 ± 0%  -0.21%  (p=0.001 n=10+10)
Insert100_Cockroach-4      2.41k ± 0%     2.41k ± 0%    ~      (p=0.059 n=9+10)
Insert1000_Cockroach-4     21.4k ± 0%     21.4k ± 0%    ~     (p=0.753 n=10+10)
Update1_Cockroach-4          636 ± 0%       635 ± 0%  -0.16%  (p=0.000 n=10+10)
Update10_Cockroach-4       1.02k ± 0%     1.02k ± 0%  -0.09%   (p=0.000 n=8+10)
Update100_Cockroach-4      4.54k ± 0%     4.54k ± 0%  -0.02%    (p=0.000 n=9+9)
Update1000_Cockroach-4     37.4k ± 0%     37.4k ± 0%    ~      (p=0.704 n=8+10)
Delete1_Cockroach-4          518 ± 0%       517 ± 0%  -0.19%    (p=0.000 n=9+9)
Delete10_Cockroach-4         772 ± 0%       771 ± 0%  -0.13%   (p=0.000 n=9+10)
Delete100_Cockroach-4      3.14k ± 0%     3.14k ± 0%  -0.08%   (p=0.000 n=10+9)
Delete1000_Cockroach-4     26.7k ± 1%     26.8k ± 0%    ~     (p=0.912 n=10+10)
```

cc. @seiflotfy

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6931)
<!-- Reviewable:end -->