@dt 2016-05-26
Reviewed 33 of 33 files at r1, 3 of 3 files at r2.
Review status: :shipit: all files reviewed at latest revision, all discussions resolved, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6931)*
<!-- Sent from Reviewable.io -->