@RaduBerinde 2016-05-26
Hi Nate,

We are passing a pointer to the planner.evalCtx to various eval functions, but these could run in parallel (e.g. for different pg clients). Won't they step on each other toes when modifying a common thing (e.g. like tmpDec)?

BTW I'm a bit skeptical on tmpDec, Are you sure this saves us anything (vs using local a `inf.Dec` variable in the functions that use it)?