@nvanbenschoten 2016-05-26
The `planner.evalCtx` is scoped to a `Session`, which is never used concurrently. If this wasn't the case, we would need to make a number of changes to make sure the `Session` and `planner` structures were thread-safe.

WRT the `tmpDec`, I'm not positive. I'll play around with it and get back to you.