@RaduBerinde 2016-05-26
Ah, right, that makes sense. LGTM!

(and the `tmpDec` thing doesn't have to hold up this PR)